services/ml-models/ffhq-wrinkle/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
opencv-python-headless==4.8.1.78
